+++ nodalia-cards.manifest.js
@@ -1,11 +1,11 @@
 export default {
-  "pkgVersion": "2.2.1-alpha.2",
-  "contentSha256_12": "910624b98f08",
+  "pkgVersion": "2.2.1-alpha.3",
+  "contentSha256_12": "d7a75ba5c659",
   "file": "nodalia-cards.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "editorSha256_12": "fcf4f0088891"
 };
-export const pkgVersion = "2.2.1-alpha.2";
-export const contentSha256_12 = "910624b98f08";
+export const pkgVersion = "2.2.1-alpha.3";
+export const contentSha256_12 = "d7a75ba5c659";
 export const file = "nodalia-cards.js";